requirements.txt
python-telegram-bot==20.7
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.40.1
torch>=2.0.0
pillow==10.2.0
numpy>=1.24.0